refactor: HumanChatPage 소켓관련 훅으로 분리
- 로직과 UI를 분리하기 위해
